refactor AccountView
