+++ PlantApp/Views/Account View/AccountViewModel.swift
@@ -1,0 +1,17 @@
+//
+//  AccountViewModel.swift
+//  PlantApp
+//
+//  Created by Aeliana Shen on 3/21/24.
+//
+
+import Foundation
+import SwiftUI
+
+class AccountViewModel: ObservableObject {
+    @AppStorage("onboarding") var isOnboardingViewActive: Bool = true
+    
+    func toggleOnboardingState() {
+        isOnboardingViewActive.toggle()
+    }
+}
